+++ tests/unit/test_solvers/test_casadi_solver.py
@@ -6,6 +6,21 @@ from scipy.sparse import eye
 
 
 class TestCasadiSolver:
+    def test_no_sensitivities_error(self):
+        model = pybamm.lithium_ion.SPM()
+        parameters = model.default_parameter_values
+        parameters["Current function [A]"] = "[input]"
+        sim = pybamm.Simulation(
+            model, solver=pybamm.CasadiSolver(), parameter_values=parameters
+        )
+        with pytest.raises(
+            NotImplementedError,
+            match="Sensitivity analysis is not implemented",
+        ):
+            sim.solve(
+                [0, 1], inputs={"Current function [A]": 1}, calculate_sensitivities=True
+            )
+
     def test_bad_mode(self):
         with pytest.raises(ValueError, match="invalid mode"):
             pybamm.CasadiSolver(mode="bad mode")